Add version update script
