+++ .github/workflows/check-ts.yml
@@ -35,4 +35,4 @@ jobs:
       working-directory: js-pkg/sdk/
 
     - name: Check that workspace package versions match
-      run: npx ts-node .github/workflows/check.ts
+      run: npx ts-node scripts/set-package-versions.ts
